+++ lib/localized_value.dart
@@ -1,0 +1,9 @@
+import 'dart:ui';
+
+/// Model that stores the locale and it's value
+class LocalizedValue {
+  Locale locale;
+  String localeValue;
+
+  LocalizedValue(this.locale, this.localeValue);
+}