+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+- Initial implementation of this package
+- Added ability to load a CSV file
+- Added ability to load one file per localizer
+- Added ability to inject Localizer
+- Added ability to set a fallback locale